refactor(mobile): modules structure
