+++ apps/mobile/src/modules/Home/HomeScreen.tsx
@@ -1,0 +1,5 @@
+import { MapView } from './components';
+
+export function HomeScreen() {
+  return <MapView />;
+}
