+++ apps/mobile/src/modules/Home/components/index.ts
@@ -1,1 +1,1 @@
-export * from './HomeScreen';
+export * from './MapView';
