+++ apps/mobile/src/modules/Home/index.ts
@@ -1,1 +1,3 @@
-export * from './components';
+import { HomeScreen } from './HomeScreen';
+
+export default HomeScreen;
