+++ apps/mobile/src/modules/Onboard/OnboardScreen.tsx
@@ -1,23 +1,11 @@
 import { Flex } from '@kavout/react-native';
 import { useNavigation, useRoute } from '@react-navigation/native';
 import { useRef } from 'react';
-import { StyleSheet } from 'react-native';
 import PagerView from 'react-native-pager-view';
 
 import { RouteParams } from '@/core/routes/types';
 
-import { ExplanationView } from './ExplanationView';
-import { LocationPermissionView } from './LocationPermissionView';
-
-const styles = StyleSheet.create({
-  viewPager: {
-    flex: 1,
-  },
-  page: {
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
+import { ExplanationView, LocationPermissionView } from './components';
 
 export function OnboardScreen() {
   const { params } = useRoute<RouteParams<'Onboard'>>();
@@ -31,7 +19,7 @@ export function OnboardScreen() {
 
   return (
     <Flex flex={1}>
-      <PagerView ref={pagerViewRef} style={styles.viewPager} initialPage={params.page} scrollEnabled={false}>
+      <PagerView ref={pagerViewRef} style={{ flex: 1 }} initialPage={params.page} scrollEnabled={false}>
         <ExplanationView onNext={() => setPage(1)} key="1" />
         <LocationPermissionView key="2" onNext={() => navigation.navigate('Home')} />
       </PagerView>
