+++ apps/mobile/src/modules/Onboard/components/index.ts
@@ -1,1 +1,2 @@
-export * from './OnboardScreen';
+export * from './ExplanationView';
+export * from './LocationPermissionView';
