+++ apps/mobile/src/modules/Onboard/index.ts
@@ -1,2 +1,4 @@
-export * from './components';
+import { OnboardScreen } from './OnboardScreen';
+
 export * from './hooks';
+export default OnboardScreen;
